Added Checks to upload offset

+++ UploadService/Models/PatchVideosModels/PatchVideosHeaders.go
@@ -3,5 +3,5 @@ package models
 type PatchVideosHeaders struct {
 	ContentType   string
 	ContentLength int
-	UploadOffset  int
+	UploadOffset  int64
 }
